refactor: README 파일의 요구사항에 맞춰 작동 과정 수정
기존: 모든 요일의 카테고리 구하고 각 코치마다 모든 요일의 메뉴 추천 -> 수정: 각 요일의 카테고리 구하고 각 코치마다 각 요일의 메뉴 추천
